fixed notification icon not showing in release mode

+++ lib/app/modules/playlist/widgets/playlist_tile.dart
@@ -1,4 +1,5 @@
 import 'package:flutter/material.dart';
+import 'package:music_player/app/widgets/dialog_confirm.dart';
 
 import '../../../data/models/playlist.dart';
 import '../../../widgets/mypopupmenu.dart';
@@ -27,18 +28,30 @@ class PlaylistTile extends StatelessWidget {
       trailing: showMenu
           ? MyPopupMenu(
               items: [
+                  MyPopupItem(id: 0, title: 'Rename', icon: Icons.edit),
                   MyPopupItem(
-                      id: 0, title: 'Delete', icon: Icons.delete_rounded),
-                  MyPopupItem(id: 1, title: 'Rename', icon: Icons.edit)
+                      id: 1, title: 'Delete', icon: Icons.delete_rounded)
                 ],
               onItemSelected: (id) {
                 switch (id) {
                   case 0:
-                    playlist.delete();
-                    break;
-                  case 1:
                     PlaylistHelper().renamePlaylist(context, playlist);
                     break;
+                  case 1:
+                    showDialog(
+                        context: context,
+                        builder: (context) => ConfirmDialog(
+                              title: 'Confirm',
+                              text: 'Are you sure to delete this playlist ?',
+                              conformText: 'Delete',
+                              onConfirm: () {
+                                playlist.delete();
+                                Navigator.pop(context);
+                              },
+                            ));
+
+                    break;
+
                   default:
                 }
               })
